chore: add hostname and domain to NeuronDescriptor
This commit adds the hostname and domain to the `NeuronDescriptor`. This
information is used to uniquely identify neurons and to group them in
the control plane. The hostname is determined by checking the `HOSTNAME`
environment variable or by reading the `/etc/hostname` file. The domain
is determined by parsing the `/etc/resolv.conf` file.